// src/api/model/organisation.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';

export interface Organisation {
    id?: string;
    name?: string;
    email?: string;
    countryCode?: string;
    contactAddress?: Address;
    contactNumber?: string;
}